Remove those triple backticks on line 8 and fix the indentation at the beginning of the file
The indentation is broken. After class GameEngine { there should be proper indentation for the constructor. The code block is also interrupted by those triple backticks on line 8.

Remove those triple backticks (```) on line 8 and fix the indentation at the beginning of the file. That should resolve the syntax error!
The error about the $ was misleading - JavaScript was getting confused by the malformed class structure.